init global integration test
