+++ vip-api/src/test/java/fr/insalyon/creatis/vip/api/rest/itest/processing/ExecutionControllerIT.java
@@ -110,10 +110,10 @@ public class ExecutionControllerIT extends BaseVIPSpringIT {
         when(workflowBusiness.getSimulation(simulation1.getID(), true))
                 .thenReturn(simulation1);
         when(workflowBusiness.getInputData(
-                 eq(simulation1.getID()), eq(baseUser1.getFolder()), any()))
+                 eq(simulation1.getID()), eq(baseUser1.getFolder())))
             .thenReturn(simulation1InData);
         when(workflowBusiness.getOutputData(
-                 eq(simulation1.getID()), eq(baseUser1.getFolder()), any()))
+                 eq(simulation1.getID()), eq(baseUser1.getFolder())))
             .thenReturn(simulation1OutData);
         mockMvc.perform(
                 get("/rest/executions/" + simulation1.getID()).with(baseUser1()))
@@ -132,10 +132,10 @@ public class ExecutionControllerIT extends BaseVIPSpringIT {
         when(workflowBusiness.getSimulation(simulation2.getID(), true))
                 .thenReturn(simulation2);
         when(workflowBusiness.getInputData(
-                 eq(simulation2.getID()), eq(baseUser1.getFolder()), any()))
+                 eq(simulation2.getID()), eq(baseUser1.getFolder())))
             .thenReturn(simulation2InData);
         when(workflowBusiness.getOutputData(
-                 eq(simulation2.getID()), eq(baseUser1.getFolder()), any()))
+                 eq(simulation2.getID()), eq(baseUser1.getFolder())))
             .thenReturn(simulation2OutData);
         mockMvc.perform(
                 get("/rest/executions/" + simulation2.getID()).with(baseUser1()))
@@ -171,10 +171,10 @@ public class ExecutionControllerIT extends BaseVIPSpringIT {
         when(workflowBusiness.getSimulation(simulation1.getID()))
                 .thenReturn(simulation1).thenThrow(new RuntimeException());
         when(workflowBusiness.getInputData(
-                 eq(simulation1.getID()), eq(baseUser1.getFolder()), any()))
+                 eq(simulation1.getID()), eq(baseUser1.getFolder())))
             .thenReturn(simulation1InData);
         when(workflowBusiness.getOutputData(
-                 eq(simulation1.getID()), eq(baseUser1.getFolder()), any()))
+                 eq(simulation1.getID()), eq(baseUser1.getFolder())))
             .thenReturn(simulation1OutData);
         mockMvc.perform(
                 put("/rest/executions/" + simulation1.getID())
@@ -202,19 +202,19 @@ public class ExecutionControllerIT extends BaseVIPSpringIT {
         when(workflowBusiness.launch(
                  eq(baseUser1), anyList(), anyMap(), eq(app1.getName()),
                  eq(version42.getVersion()), eq(class1.getName()),
-                 eq(execution1.getName()), any()))
+                 eq(execution1.getName())))
                 .thenReturn(execution1.getIdentifier());
         // configure returne execution
         when(workflowBusiness.getSimulation(execution1.getIdentifier(), true))
                 .thenReturn(simulation1);
         when(workflowBusiness.getInputData(
-                 eq(simulation1.getID()), eq(baseUser1.getFolder()), any()))
+                 eq(simulation1.getID()), eq(baseUser1.getFolder())))
             .thenReturn(simulation1InData);
         when(workflowBusiness.getOutputData(
-                 eq(simulation1.getID()), eq(baseUser1.getFolder()), any()))
+                 eq(simulation1.getID()), eq(baseUser1.getFolder())))
             .thenReturn(simulation1OutData);
         // misc config
-        when(configurationBusiness.getUserGroups(eq(baseUser1.getEmail()), any()))
+        when(configurationBusiness.getUserGroups(eq(baseUser1.getEmail())))
                 .thenReturn(new HashMap<>());
         mockMvc.perform(
                 post("/rest/executions").contentType("application/json")
@@ -231,7 +231,7 @@ public class ExecutionControllerIT extends BaseVIPSpringIT {
         verify(workflowBusiness).launch(
             eq(baseUser1), anyList(), inputCaptor.capture(), eq(app1.getName()),
             eq(version42.getVersion()), eq(class1.getName()),
-            eq(execution1.getName()), any());
+            eq(execution1.getName()));
         assertEquals(inputCaptor.getValue().size(), 2);
         MatcherAssert.<Map<?, ?>>assertThat(inputCaptor.getValue(), allOf(
                 hasEntry("param 1", "test text"),
@@ -257,14 +257,14 @@ public class ExecutionControllerIT extends BaseVIPSpringIT {
         when(workflowBusiness.getSimulation(simulation2.getID()))
                 .thenReturn(simulation2);
         when(workflowBusiness.getOutputData(
-                 eq(simulation2.getID()), eq(baseUser1.getFolder()), any()))
+                 eq(simulation2.getID()), eq(baseUser1.getFolder())))
             .thenReturn(simulation2OutData);
         String resultPath = simulation2OutData.get(0).getPath();
-        when(lfcBusiness.exists(eq(baseUser1), eq(resultPath), any()))
+        when(lfcBusiness.exists(eq(baseUser1), eq(resultPath)))
                 .thenReturn(true);
-        when(lfcBusiness.listDir(eq(baseUser1), eq(resultPath), eq(true), any()))
+        when(lfcBusiness.listDir(eq(baseUser1), eq(resultPath), eq(true)))
                 .thenReturn(Collections.singletonList(PathTestUtils.testFile1));
-        when(transferPoolBusiness.downloadFile(any(), any(), any()))
+        when(transferPoolBusiness.downloadFile(any(), any()))
                 .thenThrow(new RuntimeException());
 
         mockMvc.perform(
