+++ vip-application/src/test/java/fr/insalyon/creatis/vip/application/integrationtest/ParserIT.java
@@ -1,5 +1,6 @@
 package fr.insalyon.creatis.vip.application.integrationtest;
 
+import fr.insalyon.creatis.vip.application.client.view.system.application.ApplicationRecord;
 import fr.insalyon.creatis.vip.application.server.business.WorkflowBusiness;
 import fr.insalyon.creatis.vip.core.client.bean.Group;
 import fr.insalyon.creatis.vip.core.integrationtest.database.BaseSpringIT;
@@ -9,8 +10,13 @@ import fr.insalyon.creatis.vip.core.server.dao.GroupDAO;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.ApplicationContext;
+import org.springframework.core.io.Resource;
 
+import java.io.IOException;
 import java.util.Map;
+
+import static org.mockito.Mockito.*;
 
 /*
  To test InputM2Parser, but more importantly to test the prototype injections
@@ -24,8 +30,11 @@ public class ParserIT extends BaseSpringIT {
     @Autowired
     private GroupDAO groupDAO;
 
+    @Autowired
+    private ApplicationContext applicationContext;
+
     @Test
-    public void testInputM2Parsing() throws BusinessException, DAOException {
+    public void testInputM2Parsing() throws BusinessException, DAOException, IOException {
         // test data
         String simulationId = "input_m2_parsing_it";
         String currentUserFolder = "test_user_1";
@@ -33,6 +42,8 @@ public class ParserIT extends BaseSpringIT {
 
         // test configuration
         groupDAO.add(testGroup);
+        Resource testWorkflowPath = applicationContext.getResource("classpath:test_workflow_path");
+        when(server.getWorkflowsPath()).thenReturn(testWorkflowPath.getFile().getAbsolutePath());
 
         // do test
         Map<String, String> res = workflowBusiness.relaunch(simulationId, currentUserFolder);
