+++ vip-core/src/test/java/fr/insalyon/creatis/vip/core/integrationtest/SpingTestConfig.java
@@ -3,6 +3,7 @@ package fr.insalyon.creatis.vip.core.integrationtest;
 import fr.insalyon.creatis.grida.client.GRIDAClient;
 import fr.insalyon.creatis.grida.client.GRIDAPoolClient;
 import fr.insalyon.creatis.vip.core.server.business.EmailBusiness;
+import fr.insalyon.creatis.vip.core.server.business.proxy.ProxyClient;
 import org.mockito.Mockito;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -21,4 +22,7 @@ public class SpingTestConfig {
 
     @Bean @Primary
     public EmailBusiness testEmailBusiness() {return Mockito.mock(EmailBusiness.class);}
+
+    @Bean @Primary
+    public ProxyClient testProxyClient() {return Mockito.mock(ProxyClient.class);}
 }
