+++ vip-portal/src/test/resources/vip-api.conf
@@ -1,0 +1,31 @@
+# comma separated list of authorized domains
+# domain example : http://brain-perfusion.creatis.insa-lyon.fr
+cors.autorized-domains=
+
+# General info stuff
+carmin.platform.name=VIP_TEST
+carmin.platform.description=Test instance of Virtual Imaging Platform
+carmin.platform.supported_transfer_protocols=HTTPS
+carmin.platform.supported_modules=PROCESSING
+carmin.platform.default_limit_list_execution=500
+carmin.platform.unsupported_methods=playExecution
+carmin.platform.supported_API_Version=0.3
+carmin.platform.email=test@test.tst
+carmin.platform.error_codes_and_message=\
+  42:Test error 1,\
+  4242:Test error 2
+
+# Authentication stuff
+carmin.authentication.apikey.header.name=apikey_test
+carmin.authentication.apikey.generate_new_key_on_each_authentication=false
+
+# DATA stuff
+carmin.data.path.download=/path
+carmin.data.mime_type.directory=text/directory
+carmin.data.mime_type.default=application/octet-stream
+carmin.data.download.retry=2
+carmin.data.download.timeout=45
+carmin.data.max_size=15000000
+
+# Pipeline stuff
+carmin.processing.pipelines.whitelist=
